AV: move Boards Page components into 1 folder
