+++ Frontend/user/src/components/BoardsPage/BoardsPageCardTemplate.tsx
