+++ Frontend/user/src/components/BoardsPage/BoardsPageRowTemplate.tsx
